+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # 2.x branch
 ## 2.30 branch
 ### 2.30.0
+* The topbar element will use the `TopbarHelper` from APP to build links, if
+    that helper exists. Otherwise it will use the helper provided by MeCms, with
+    the helper of any other plugin;
 * the sitemap classes have been moved into the `MeCms\Utility\Sitemap` namespace.
     Added the `SitemapBase` abstract class, that provides the `parse()` method
     and which now needs to be extended by the `Sitemap` class.